update_engine: Remove problematic deflate caches.
This patch calls puffin::RemoveDeflatesWithBadDistanceCaches() to ignore certain
deflates that older puffin clients fail to properly process.

Unfortunately this adds another round of puffin call which will add a few
seconds to paygen, but we can't get out of it at this point yet. In the future
we can increase a minor version for it.

Bug: crbug.com/915559
Test: generate delta update from payloads with problematic deflates
Change-Id: I77416d4005b5ad2fbb47c129541c811b6a380f60