Abort the payload generation on extract_image errors.
extract_image, in particular extract_image_brillo, spawns multiple
processes to handle the unzipping and unsparsing works in parallel.
However, `wait` doesn't abort the execution on child process errors.

This CL tracks the PIDs of the worker processes. It then waits for each
individual PID and aborts appropriately. No-op for the CrOS path.

Bug: 129773355
Test: Inject error into simg2img. brillo_update_payload aborts.
Change-Id: Id87aa5216fee7e9db132b7f7947742d2ac11b2bc