Set the block device to be writable before writing verity data
The block device is set to be writable when the delta performer opens
the partition to perfrom writes. But if the update interrupts and the
device reboots during the verity data calculation, update engine will
consider the write actions done on resumes and skip setting the writable
flag. As a result, the resumed update will fail to write the hashtree
with EPERM.

A possible fix is to explicltly set the flag when the FilesystemVerifierAction
wants to write verity data.

Bug: 131232680
Test: Interrupt the update during verity writing, reboot and check the
update resumes correctly

Change-Id: I1007414139e4ae5be3fc60ec73b18264d4baa60e