Support `atest update_engine_unittests`.
We need a few changes to the building rules to support running the test
with atest. In particular, we need a customized test config to filter
out the helper executables. The change doesn't affect the traditionally
way of running tests with `adb sync data` + `adb shell
/d/n/u/update_engine_unittests`.

This is one of the prerequisites to run update_engine_unittests with
TEST_MAPPING [1]. It still requires additional works to enable the test
in pre/post-submit though (currently it's failing on GCE devices due to
missing shared libraries).

[1] https://source.android.com/compatibility/tests/development/test-mapping

Fixes: 131438455
Test: `atest update_engine_unittests` on blueline.
Test: TreeHugger; check that the existing GCL config'd test keeps working.
Change-Id: If602d69d84e0d926c579324d988c94b57240b113