Skip an async callback function when the UpdateBootFlagsAction object is destroyed
As the callback to MarkBootSuccessfulAsync, UpdateBootFlagsAction's
member function CompleteUpdateBootFlags() can still be called even after
the current UpdateBootFlagsAction object get destroyed by the action
processor. We want to set a static flag in TerminateProcessing() and check
its value before executing the callback function.

An alternative way is to save and propagate the TaskId when scheduling the
task in MarkBootSuccessfulAsync, and cancel the task in UpdateBootFlagsAction's
TerminateProcessing().

Bug: 123720545
Test: No longer hit the CHECK after injecting StopProcessing.
Change-Id: I98d2cc7b94d4059fb897b89932969b61936e8c2e