update_engine: Bug fix for UpdateEngineStatus print
Exempt-From-Owner-Approval: The auto_updater.py depends on the status of
update_engine from the printout of `--status`. It finds the key
`CURRENT_OP`, but the CL in chromium:1715978 set that to
`CURRENT_OPERAITON`. Is required for `cros flash` to work properly.

Revert CURRENT_OPERATION to previous CURRENT_OP.

Output now:
  [0801/095624.227871:INFO:update_engine_client.cc(490)] Querying Update Engine status...
  CURRENT_OP=UPDATE_STATUS_IDLE
  IS_INSTALL=false
  LAST_CHECKED_TIME=0
  NEW_SIZE=0
  NEW_VERSION=0.0.0.0
  PROGRESS=0.0

BUG=chromium:871340
TEST=FEATURES="test" emerge-$B update_engine
TEST=update_engine_client --status

Change-Id: I23142dab51894adc2aeeb06f0459c74287b1639b