Stop using unzip -p when extracting images
brillo_update_payload uses `unzip -p` to extract images to a named temp
file. But the command doesn't generate an error when the disk is full.
This cl creates a helper function and switches to use `unzip -d` to extract
the image; and then move the extracted image to the intended name.

Bug: 138725843
Test: generate and verify a payload, check the extract error when disk is full
Change-Id: I71c2d07de3c1c826f2e07fcc6497437c4051944f