update_engine: Fix field assignment order, for compiler warning.
The latest llvm compiler version (r370808), to which we are about to
upgrade Chrome OS, introduces a new warning, -Wreorder-init-list, which
complains if struct field assignments at initialization occur in a
different order than their declartion order.  This CL fixes a build
error in the ChromeOS UnitTest stage that is uncovered by this warning.

BUG=chromium:1006866
TEST=compilation works with this fix, with new compiler

Change-Id: Id48ae7b24345d960313eaa03ce58378065a9316e
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/1818788
Tested-by: Caroline Tice <cmtice@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>