update_engine: remove an unreachable return
We're trying to turn -Wunreachable-code-return on by default, and this
unreachable `return` statement was flagged. The code above this is
pretty clearly an infloop, so this return statement doesn't seem super
useful. If someone adds a `break` or similar later, the compiler will
complain (in -Werror form) about a missing `return` statement.

BUG=chromium:1066998
TEST=CQ

Change-Id: I9c87e256fa0c850bcb8cec17cbe694fc6bb0ef66
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/2133017
Reviewed-by: Sen Jiang <senj@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>