update_engine: Change parameters of AttemptInstall() DBus
Currently, the update_engine is using the dlcservice's protobuf as
its own input. This is bad API because now anyone dependent on the
update_engine's DBus, needs to be dependent on the dlcservice too. This
CL changes that so we only pass the arguments we want (DLC DIs and the
omaha url).

BUG=chromium:1071654
TEST=cros_workon_make --board reef --test update_engine

Cq-Depend: chromium:2157670
Change-Id: Id7f7a86d8b3e3194d4d7697a8ad26ed4bcc4ba2c
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/2163442
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>